requirements.txt
capstone
python-magic
pefile
pyelftools
